linker: Cleanup
