+++ src/core/tls.h
@@ -3,8 +3,14 @@
 
 #pragma once
 
+#include "common/types.h"
+
 namespace Core {
 
+/// Installs a host exception handler to handle guest TLS access.
 void InstallTlsHandler();
+
+/// Patches any instructions that access TLS to trigger the exception handler.
+void PatchTLS(u64 segment_addr, u64 segment_size);
 
 } // namespace Core
